test: run the free-socket guard cases on Node.js and Bun from one node:test file
Move the cases into node-http-agent-free-socket.node.mts (node:test +
node:assert, TLS cert from fixtures/) and run it from node-http.test.ts
under both `node --test` and `bun test`, like node-http-connect. The
guard cases describe Node from v26.4.0 (nodejs/node 179ddaedfb), so they
skip on an older Node. The queued-request case is Bun-only: Node checks a
freed socket for buffered bytes only on its way into the pool.
